Fix buffer overflow

+++ src/source/Ice/Network.c
@@ -336,7 +336,7 @@ CleanUp:
 
 BOOL isIpAddr(PCHAR hostname, UINT16 length)
 {
-    UINT16 offset = 0;
+    UINT32 offset = 0;
     UINT32 ip_1, ip_2, ip_3, ip_4, ip_5, ip_6, ip_7, ip_8;
 
     if (hostname == NULL) {
